refactor: move Meshes to extension
